ias-shell: Added a set_soc request for remote display
An ias_hmi client app can inform us for any surface that may
not be displayed locally and is meant for another SoC by setting
a flag in a request called set_soc. If bit 0 is set, then this
surface is meant for local display, if bit 1 is set, it is for
SoC #1 and so on.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>