ias-hmi: Added client_created request
This request lets an HMI know when a client is about to connect
to the compositor. Usually, the HMI is the launcher of these
client apps and can get the pid of the client when it does a fork
call.
In remote display use cases, the compositor may want to hide certain
Wayland outputs from some clients that are meant only for remote
SoCs. However, in order to hide these Wayland outputs, the compositor
must know about this client even before it connects and starts
obtaining the registry info so that the compositor can filter it out.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>